:panda_face: Maps

+++ database/src/main/java/mx/com/charlyescaz/database/models/WeatherBD.kt
@@ -11,5 +11,5 @@ data class WeatherBD(
         val main: String?,
         val description: String?,
         val icon: String?,
-        var fkIdTest: Int?
+        var fkIdTest: Long?
 )
